create apiclient copies of repositories in data access layer + add to extension method
