+++ GeoMastery.BlazorWASM/Extensions/DependencyInjection.cs
@@ -1,4 +1,5 @@
-﻿using GeoMastery.BlazorWASM.Repositories;
+﻿using GeoMastery.BlazorWASM.Repositories.ApiClient;
+using GeoMastery.BlazorWASM.Repositories.Local;
 using GeoMastery.Persistence.Repositories.v1;
 
 namespace GeoMastery.BlazorWASM.Extensions;
@@ -14,6 +15,8 @@ public static class DependencyInjection
 
     public static void UseRemoteRepositories(this IServiceCollection services)
     {
-
+        services.AddScoped<ICountryRepository, CountryApiClient>();
+        services.AddScoped<IContinentRepository, ContinentApiClient>();
+        services.AddScoped<IRegionRepository, RegionApiClient>();
     }
 }
